Utils.java Static Helper for Marquee
