+++ app/src/main/java/com/amaze/filemanager/adapters/CompressedExplorerAdapter.java
@@ -220,7 +220,8 @@ public class CompressedExplorerAdapter extends RecyclerView.Adapter<CompressedIt
 
     boolean enableMarquee =
         sharedPrefs.getBoolean(PreferencesConstants.PREFERENCE_ENABLE_MARQUEE_FILENAME, true);
-    AnimUtils.configureTitleMarquee(holder.firstLine, enableMarquee);
+    holder.txtTitle.setEllipsize(
+        enableMarquee ? TextUtils.TruncateAt.MARQUEE : TextUtils.TruncateAt.MIDDLE);
 
     final CompressedObjectParcelable rowItem = items.get(position);
     GradientDrawable gradientDrawable = (GradientDrawable) holder.genericIcon.getBackground();
